component(button): Created button components
Adapted code styling to better suit the project.
Tried to make it feel more like golang to align with what we know
and like.

Created 3 distinct fully featured button types with hover and action
states.

Buttons also adhere to HTML5 semantics by automatically rendering
as anchor tags for links and buttons for all other actions which are not
navigation.